@@ -11,34 +11,34 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^8.2.14",
+    "@angular/animations": "^9.0.2",
     "@angular/cdk": "^8.2.3",
-    "@angular/common": "^8.2.14",
-    "@angular/compiler": "^8.2.14",
-    "@angular/core": "^8.2.14",
+    "@angular/common": "^9.0.2",
+    "@angular/compiler": "^9.0.2",
+    "@angular/core": "^9.0.2",
     "@angular/flex-layout": "^8.0.0-beta.27",
-    "@angular/forms": "^8.2.14",
+    "@angular/forms": "^9.0.2",
     "@angular/material": "^8.2.3",
-    "@angular/platform-browser": "^8.2.14",
-    "@angular/platform-browser-dynamic": "^8.2.14",
-    "@angular/router": "^8.2.14",
+    "@angular/platform-browser": "^9.0.2",
+    "@angular/platform-browser-dynamic": "^9.0.2",
+    "@angular/router": "^9.0.2",
     "@waiting/base64": "^4.2.9",
     "angular-oauth2-oidc": "^8.0.4",
     "hammerjs": "^2.0.8",
     "rxjs": "~6.5.4",
     "rxjs-compat": "^6.5.4",
     "tslib": "^1.10.0",
-    "zone.js": "~0.9.1"
+    "zone.js": "~0.10.2"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^0.803.22",
-    "@angular/cli": "^8.3.22",
-    "@angular/compiler-cli": "^8.2.14",
-    "@angular/language-service": "^8.2.14",
+    "@angular-devkit/build-angular": "~0.900.3",
+    "@angular/cli": "^9.0.3",
+    "@angular/compiler-cli": "^9.0.2",
+    "@angular/language-service": "^9.0.2",
     "@types/jasmine": "~3.5.0",
     "@types/jasminewd2": "^2.0.8",
-    "@types/node": "~13.1.7",
-    "codelyzer": "^5.2.1",
+    "@types/node": "^12.11.1",
+    "codelyzer": "^5.1.2",
     "jasmine-core": "~3.5.0",
     "jasmine-spec-reporter": "~4.2.1",
     "karma": "~4.4.1",
@@ -49,6 +49,6 @@
     "protractor": "~5.4.2",
     "ts-node": "^8.6.2",
     "tslint": "~5.20.1",
-    "typescript": ">=3.1 < 3.6"
+    "typescript": "3.7.5"
   }
 }
